+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Game notification
 - Commit-style transform construction
 - Game history
+- Plugin System
 
 ### Updated
 
@@ -30,6 +31,8 @@
 - Added static method `Menu.prompt` as an alias of `new Menu`
 - Use `game.config.defaultTextColor` to set the default text color
 - Added `createPersistent` as a shortcut for initializing a persistent
+- Added game hooks to `Game`
+- Added `Character.config.color` to set the color of the character's name tag
 
 ### Fixed